Add and Get Survivor with Items fully working

+++ src/main/java/com/backendrecruitmenttest/ZSSN/repository/SurvivorRepository.java
@@ -4,5 +4,5 @@ package com.backendrecruitmenttest.ZSSN.repository;
 import com.backendrecruitmenttest.ZSSN.entity.Survivor;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SurvivorRepository extends CrudRepository<Survivor, Integer> {
+public interface SurvivorRepository extends CrudRepository<Survivor, Long> {
 }
